etl_ugc/requirements.txt
clickhouse_driver==0.2.4
pydantic==1.10.2
backoff==2.2.1
kafka-python==2.0.2